Fix: MemberAggregationService를 AggregationService로 리팩토링
