+++ src/main/java/com/meme/ala/domain/aggregation/service/AggregationService.java
@@ -5,4 +5,6 @@ import com.meme.ala.domain.member.model.entity.Member;
 
 public interface AggregationService {
     public Aggregation findByMember(Member member);
+
+    public void initAggregation(Member member);
 }
