--- src/main/java/com/meme/ala/domain/member/service/MemberAggregationService.java
@@ -1,7 +1,0 @@
-package com.meme.ala.domain.member.service;
-
-import com.meme.ala.domain.member.model.entity.Member;
-
-public interface MemberAggregationService {
-    public void initAggregation(Member member);
-}
